Combine stepper interface classes.
