+++ v2/src/shared/StepperInterface.cc
@@ -18,10 +18,27 @@
 #include "StepperInterface.hh"
 #include "Eeprom.hh"
 #include "EepromMap.hh"
+#include "Configuration.hh"
+
+StepperInterface::StepperInterface(const Pin& dir,
+                                   const Pin& step,
+                                   const Pin& enable,
+                                   const Pin& max,
+                                   const Pin& min,
+                                   const uint16_t eeprom_base_in) :
+    dir_pin(dir),
+    step_pin(step),
+    enable_pin(enable),
+    max_pin(max),
+    min_pin(min),
+    invert_endstops(true),
+    invert_axis(false),
+    eeprom_base(eeprom_base_in) {
+}
 
 void StepperInterface::setDirection(bool forward) {
 	if (invert_axis) forward = !forward;
-	dir_pin.setValue(forward);
+        dir_pin.setValue(forward);
 }
 
 void StepperInterface::step(bool value) {
@@ -34,12 +51,14 @@ void StepperInterface::setEnabled(bool enabled) {
 }
 
 bool StepperInterface::isAtMaximum() {
+        if (max_pin.isNull()) return false;
 	bool v = max_pin.getValue();
 	if (invert_endstops) v = !v;
 	return v;
 }
 
 bool StepperInterface::isAtMinimum() {
+        if (min_pin.isNull()) return false;
 	bool v = min_pin.getValue();
 	if (invert_endstops) v = !v;
 	return v;
@@ -50,17 +69,22 @@ void StepperInterface::init(uint8_t idx) {
 	step_pin.setDirection(true);
 	enable_pin.setValue(true);
 	enable_pin.setDirection(true);
-	max_pin.setDirection(false);
-	min_pin.setDirection(false);
 	// get inversion characteristics
-	uint8_t axes_invert = eeprom::getEeprom8(eeprom::AXIS_INVERSION, 1<<1);
-	uint8_t endstops_invert = eeprom::getEeprom8(eeprom::ENDSTOP_INVERSION, 0);
+        uint8_t axes_invert = eeprom::getEeprom8(eeprom_base, 1<<1);
+        uint8_t endstops_invert = eeprom::getEeprom8(eeprom_base + 1, 0);
 	bool endstops_present = (endstops_invert & (1<<7)) != 0;
+
 	// If endstops are not present, then we consider them inverted, since they will
 	// always register as high (pulled up).
 	invert_endstops = !endstops_present || ((endstops_invert & (1<<idx)) != 0);
 	invert_axis = (axes_invert & (1<<idx)) != 0;
-	// pull pins up to avoid triggering when using inverted endstops
-	max_pin.setValue(invert_endstops);
-	min_pin.setValue(invert_endstops);
+        // pull pins up to avoid triggering when using inverted endstops
+        if (!max_pin.isNull()) {
+                max_pin.setDirection(false);
+                max_pin.setValue(invert_endstops);
+        }
+        if (!min_pin.isNull()) {
+                min_pin.setDirection(false);
+                min_pin.setValue(invert_endstops);
+        }
 }
